+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added outbox POST handler (FEP-ae97).
+
 ### Changed
 
 - Update `/api/v1/accounts/send_activity` to accept only signed activities.